+++ src/deepskylog/AstronomyLibrary/AstronomyLibrary.php
@@ -309,7 +309,7 @@ class AstronomyLibrary
      */
     public function getLengthOfNightPlot($timezone): string
     {
-        if (! $this->_lengthOfNightChart) {
+        if (!$this->_lengthOfNightChart) {
             $date = Carbon::now();
             $date->year($this->getDate()->year);
 
@@ -684,7 +684,7 @@ class AstronomyLibrary
             $rawImageBytes = ob_get_clean();
 
             $this->_lengthOfNightChart = "<img src='data:image/jpeg;base64,"
-                .base64_encode($rawImageBytes)."' />";
+                . base64_encode($rawImageBytes) . "' />";
         }
 
         return $this->_lengthOfNightChart;
@@ -745,5 +745,18 @@ class AstronomyLibrary
         EquatorialCoordinates $coords3
     ): Coordinate {
         return $coords1->smallestCircle($coords2, $coords3);
+    }
+
+    /**
+     * Returns the apparent place of a star.
+     *
+     * @param EquatorialCoordinates $coords The coordinates to start with
+     *
+     * @return EquatorialCoordinates The apparent place for the star
+     */
+    public function apparentPlace(
+        EquatorialCoordinates $coords
+    ): EquatorialCoordinates {
+        return $coords->apparentPlace($this->getDate(), $this->getNutation());
     }
 }